+++ src/ChartExports.js
@@ -13,16 +13,60 @@ const makeCSVExport = data => {
   return `data:text/csv;charset=utf-8;base64,${btoa(csvContent)}`;
 };
 
+const makePNGExport = imageSrc => new Promise(resolve => {
+  const canvas = document.createElement("canvas");
+
+  const size = 800;
+  const padding = 100;
+
+  canvas.width = size;
+  canvas.height = size;
+
+  const image = new Image();
+  image.onload = () => {
+    const context = canvas.getContext("2d");
+    context.fillStyle = "white";
+
+    context.fillRect(0, 0, size, size);
+    context.drawImage(
+      image, 0, 0, size, size,
+      padding, padding, size - padding * 2, size - padding * 2
+    );
+
+    resolve(canvas.toDataURL("image/png"));
+  };
+  image.src = imageSrc;
+});
+
 const makeSVGExport = svg => {
   const svgContent = new XMLSerializer().serializeToString(svg);
   return `data:image/svg+xml;base64,${btoa(svgContent)}`;
 };
 
+const ChartExport = ({ ext, href, tabIndex }) => (
+  <a
+    href={href}
+    target="_blank"
+    rel="noopener noreferrer"
+    download={`chart.${ext}`}
+    tabIndex={tabIndex}
+  >
+    Export .{ext}
+  </a>
+);
+
 class ChartExports extends PureComponent {
   constructor(props) {
     super(props);
 
-    this.state = { csvExport: "#", svgExport: "#" };
+    this.state = {
+      csvExport: "#",
+      dropdownOpen: false,
+      pngExport: "#",
+      svgExport: "#"
+    };
+
+    this.handleToggleDropdown = this.handleToggleDropdown.bind(this);
   }
 
   componentDidMount() {
@@ -31,10 +75,15 @@ class ChartExports extends PureComponent {
   }
 
   componentDidUpdate(prevProps) {
-    const { data } = this.props;
+    const { data, hovering } = this.props;
+    const { dropdownOpen } = this.state;
 
     if (data !== prevProps.data) {
       this.makeExports();
+    }
+
+    if (hovering !== prevProps.hovering && !hovering && dropdownOpen) {
+      this.setState({ dropdownOpen: false });
     }
   }
 
@@ -44,51 +93,52 @@ class ChartExports extends PureComponent {
 
   makeExports() {
     setTimeout(() => {
-      if (this.componentIsMounted) {
-        const { data, svgRef } = this.props;
-
-        this.setState({
-          csvExport: makeCSVExport(data),
-          svgExport: makeSVGExport(svgRef.current)
-        });
+      if (!this.componentIsMounted) {
+        return;
       }
+
+      const { data, svgRef } = this.props;
+
+      const csvExport = makeCSVExport(data);
+      const svgExport = makeSVGExport(svgRef.current);
+
+      makePNGExport(svgExport).then(pngExport => {
+        if (this.componentIsMounted) {
+          this.setState({ csvExport, pngExport, svgExport });
+        }
+      });
     }, 2000);
   }
 
+  handleToggleDropdown() {
+    this.setState(({ dropdownOpen }) => ({
+      dropdownOpen: !dropdownOpen
+    }));
+  };
+
   render() {
     const { hovering } = this.props;
-    const { csvExport, svgExport } = this.state;
-
-    const tabIndex = hovering ? 0 : -1;
+    const { csvExport, dropdownOpen, pngExport, svgExport } = this.state;
 
     return (
       <div className="chq-charts--export">
-        <a
-          href={csvExport}
-          target="_blank"
-          rel="noopener noreferrer"
-          download="chart.csv"
-          title="Export data"
-          aria-label="Export data"
-          tabIndex={tabIndex}
+        <button
+          type="button"
+          onClick={this.handleToggleDropdown}
+          aria-label="Open dropdown"
+          tabIndex={hovering ? 0 : -1}
         >
           <svg viewBox="0 0 1024 1024">
-            <path d="M864 160v704h-704v-704h704zM896 128h-768v768h768v-768z M384 304h384v32h-384v-32z M384 496h384v32h-384v-32z M384 688h384v32h-384v-32z M320 320c0 17.673-14.327 32-32 32s-32-14.327-32-32c0-17.673 14.327-32 32-32s32 14.327 32 32z M320 512c0 17.673-14.327 32-32 32s-32-14.327-32-32c0-17.673 14.327-32 32-32s32 14.327 32 32z M320 704c0 17.673-14.327 32-32 32s-32-14.327-32-32c0-17.673 14.327-32 32-32s32 14.327 32 32z" />
+            <path d="M227.4 608c-55 0-99.4-42.8-99.4-96 0-53 44.4-96 99.4-96 55.2 0 99.6 43 99.6 96 0 53.2-44.4 96-99.6 96z M512 608c-55 0-99.6-42.8-99.6-96 0-53 44.6-96 99.6-96s99.4 43 99.4 96c0 53.2-44.4 96-99.4 96z M796.4 608c-55 0-99.6-42.8-99.6-96 0-53 44.4-96 99.6-96 55 0 99.6 43 99.6 96 0 53.2-44.4 96-99.6 96z" />
           </svg>
-        </a>
-        <a
-          href={svgExport}
-          target="_blank"
-          rel="noopener noreferrer"
-          download="chart.svg"
-          title="Export graphic"
-          aria-label="Export graphic"
-          tabIndex={tabIndex}
-        >
-          <svg viewBox="0 0 1024 1024">
-            <path d="M835 320h-123c-64-72-84-96-109-96h-177c-25 0-44 24-109 96h-27v-32h-68v32h-27c-35 0-67 26-67 61v352c0 35 32 67 67 67h640c35 0 61-32 61-67v-352c0-35-26-61-61-61z M864 733c0 19-12 35-29 35h-640c-17 0-35-17-35-35v-352c0-16 16-29 35-29h136l10-6c8-9 15-20 22-28 23-25 39-43 51-54 9-8 12-8 12-8h177c0 0 3 0 13 9 12 11 29 33 53 60 6 7 12 14 19 21l10 6h138c18 0 29 12 29 29v352z M512 379c-94 0-171 77-171 171s77 171 171 171 171-77 171-171-77-171-171-171z M512 689c-77 0-139-62-139-139s62-139 139-139 139 62 139 139-62 139-139 139z M704 384h34v34h-34v-34z M576 550c0 35-29 64-64 64s-64-29-64-64c0-35 29-64 64-64s64 29 64 64z" />
-          </svg>
-        </a>
+        </button>
+        {dropdownOpen && (
+          <div className="chq-charts--export-dropdown">
+            <ChartExport ext="csv" href={csvExport} />
+            <ChartExport ext="png" href={pngExport} />
+            <ChartExport ext="svg" href={svgExport} />
+          </div>
+        )}
       </div>
     );
   }